+++ types/handler.go
@@ -40,14 +40,18 @@ func ChainAnteDecorators(chain ...AnteDecorator) AnteHandler {
 		return nil
 	}
 
-	// handle non-terminated decorators chain
-	if (chain[len(chain)-1] != Terminator{}) {
-		chain = append(chain, Terminator{})
+	handlerChain := make([]AnteHandler, len(chain)+1)
+	// Install the terminal AnteHandler.
+	handlerChain[len(chain)] = func(ctx Context, tx Tx, simulate bool) (Context, error) {
+		return ctx, nil
 	}
-
-	return func(ctx Context, tx Tx, simulate bool) (Context, error) {
-		return chain[0].AnteHandle(ctx, tx, simulate, ChainAnteDecorators(chain[1:]...))
+	for i := 0; i < len(chain); i++ {
+		ii := i
+		handlerChain[ii] = func(ctx Context, tx Tx, simulate bool) (Context, error) {
+			return chain[ii].AnteHandle(ctx, tx, simulate, handlerChain[ii+1])
+		}
 	}
+	return handlerChain[0]
 }
 
 // ChainPostDecorators chains PostDecorators together with each PostDecorator
@@ -63,14 +67,18 @@ func ChainPostDecorators(chain ...PostDecorator) PostHandler {
 		return nil
 	}
 
-	// handle non-terminated decorators chain
-	if (chain[len(chain)-1] != Terminator{}) {
-		chain = append(chain, Terminator{})
+	handlerChain := make([]PostHandler, len(chain)+1)
+	// Install the terminal PostHandler.
+	handlerChain[len(chain)] = func(ctx Context, tx Tx, simulate, success bool) (Context, error) {
+		return ctx, nil
 	}
-
-	return func(ctx Context, tx Tx, simulate, success bool) (Context, error) {
-		return chain[0].PostHandle(ctx, tx, simulate, success, ChainPostDecorators(chain[1:]...))
+	for i := 0; i < len(chain); i++ {
+		ii := i
+		handlerChain[ii] = func(ctx Context, tx Tx, simulate, success bool) (Context, error) {
+			return chain[ii].PostHandle(ctx, tx, simulate, success, handlerChain[ii+1])
+		}
 	}
+	return handlerChain[0]
 }
 
 // Terminator AnteDecorator will get added to the chain to simplify decorator code